move json editor e2e to separate spec file
